refactor: rename dashboard to aplikasi
